linux, scmi: add scmi configuration.
The scope:
Add the device tree to support scmi for xen and domains domd and doma.
The defconfig with the scmi options is added into folder linux/files/scmi.
It is required to have the possibility to build the product with or without the scmi feature.
Additionally, the scmi feature is added to the boot options file.

Signed-off-by: Ihor Usyk <ihor_usyk@epam.com>